feat(config): 调整镜像同步和清理配置
- 将 max_parallel_downloads 从 5 调整为 1
- 将 auto_sync_on_start 从 false 调整为 true
- 将 keep_versions 从 5 调整为 9999

feat(server): 增加自动同步功能

- 在服务器启动时检查是否需要自动同步
- 如果配置启用自动同步，则在后台启动同步进程

fix(server): 优化服务器启动端口获取逻辑

- 如果命令行未指定端口，则从配置文件中获取

fix(web): 修复 Linux 系统下 uptime 信息获取

- 修复了隐式类型转换导致的警告

+++ srcMirror/Web/Controller.php
@@ -329,10 +329,11 @@ class Controller
         if (PHP_OS_FAMILY === 'Linux') {
             $uptime = @file_get_contents('/proc/uptime');
             if ($uptime !== false) {
-                $uptime = explode(' ', $uptime)[0];
-                $days = floor($uptime / 86400);
-                $hours = floor(($uptime % 86400) / 3600);
-                $minutes = floor(($uptime % 3600) / 60);
+                $uptime = (float)explode(' ', $uptime)[0];
+                // 显式转换避免隐式转换警告
+                $days = (int)floor((float)$uptime / 86400);
+                $hours = (int)floor((float)($uptime % 86400) / 3600);
+                $minutes = (int)floor((float)($uptime % 3600) / 60);
 
                 return "{$days}天 {$hours}小时 {$minutes}分钟";
             }
